Move wf_dict serializer to flowrep_dict
